Fix unsafety in PrintableErrno.eprint_signalsafe
In situations where signal-safety is desired (e.g. after a call to fork() on a multithreaded program), PrintableErrno.eprint_signalsafe provides a way to print an error code to stderr in a signal-safe way. Unfortunately, the function allocated a vector to the heap (using malloc() internally), breaking the established contract. This commit re-writes the offending portion by calling write() many times.

Note: if writev is signal-safe, its use should be explored to reduce the amount of syscalls.